refactor: change structure of code
