+++ src/components/simple-survey.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Image, Text, TouchableOpacity, View } from 'react-native';
 import Images from '../assets/images';
 import SurveyStyles from '../styles/simple-survey.styles';
-import type { Survey, SurveyProperties } from '../api/bitlabs/types';
+import type { Survey, SurveyProperties } from '../api/bitlabs_service/types';
 import { RewardView } from '../hoc/reward-view';
 import { currencize, rounded } from '../utils/helpers';
